pyproject.toml
[project]
name = "cc-optimize"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "gepa>=0.1.0",
]

[project.optional-dependencies]
sdk = [
    "claude-agent-sdk>=0.1.0",
]
dev = [
    "pytest>=7.0",
]

[project.scripts]
cc-optimize = "cc_optimize.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
